+++ packages/eslint_config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @myworldx/eslint-config
 
+## 4.0.1
+
+### Patch Changes
+
+- Disabled no require statement
+
+## 4.0.0
+
+### Major Changes
+
+- Change import config from eslint to prettier, add prettier import plugin
+
 ## 3.2.0
 
 ### Minor Changes